inventory-service/pyproject.toml
[tool.poetry]
name = "inventory-service"
version = "0.1.0"
description = "Book Store Inventory Service"
authors = ["bookstore"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
aiokafka = "^0.11.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
